+++ src/app/_types/sharedProps/imageType.ts
@@ -1,0 +1,4 @@
+export type ImageProps = {
+  url: string
+  alt: string
+}
